+++ src/Genius/ConnectGenius.php
@@ -7,12 +7,12 @@ use Genius\Authentication\OAuth2;
 use Http\Client\Common\Plugin\AddHostPlugin;
 use Http\Client\Common\Plugin\AuthenticationPlugin;
 use Http\Client\Common\PluginClient;
+use Http\Client\Common\PluginClientFactory;
 use Http\Client\HttpClient;
-use Http\Discovery\HttpClientDiscovery;
-use Http\Discovery\UriFactoryDiscovery;
+use Http\Discovery\Psr17FactoryDiscovery;
+use Http\Discovery\Psr18ClientDiscovery;
 use Http\Message\Authentication;
-use Http\Message\UriFactory;
-use Psr\Http\Message\UriInterface;
+use Psr\Http\Message\UriFactoryInterface;
 
 class ConnectGenius
 {
@@ -21,7 +21,7 @@ class ConnectGenius
     /** @var HttpClient */
     protected $httpClient;
     
-    /** @var  UriFactory */
+    /** @var  UriFactoryInterface */
     protected $uriFactory;
     
     /** @var Authentication|OAuth2 */
@@ -44,23 +44,20 @@ class ConnectGenius
             new AuthenticationPlugin($this->authentication),
         ];
 
-        return new PluginClient(
-            $this->getHttpClient(),
-            $plugins
-        );
+        return (new PluginClientFactory())->createClient($this->getHttpClient(), $plugins);
     }
     
-    public function setUriFactory(UriInterface $uriFactory): ConnectGenius
+    public function setUriFactory(UriFactoryInterface $uriFactory): ConnectGenius
     {
         $this->uriFactory = $uriFactory;
         
         return $this;
     }
     
-    public function getUriFactory(): UriFactory
+    public function getUriFactory(): UriFactoryInterface
     {
         if ($this->uriFactory === null) {
-            $this->uriFactory = UriFactoryDiscovery::find();
+            $this->uriFactory = Psr17FactoryDiscovery::findUrlFactory();
         }
         
         return $this->uriFactory;
@@ -92,7 +89,7 @@ class ConnectGenius
     protected function getHttpClient(): HttpClient
     {
         if ($this->httpClient === null) {
-            $this->httpClient = HttpClientDiscovery::find();
+            $this->httpClient = Psr18ClientDiscovery::find();
         }
         
         return $this->httpClient;